requirements.txt
absl-py==0.13.0
addict==2.4.0
albumentations==1.0.0
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.1.6
colour==0.1.5
cycler==0.10.0
cysignals==1.11.2
decorator==4.4.2
dill==0.3.6
easydict==1.9
editdistance==0.5.3
fire==0.4.0
Flask==2.2.5
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.6.0
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.2.1
lpips==0.1.4
Markdown==3.3.4
MarkupSafe==2.1.3
matplotlib==3.4.2
matplotlib-inline==0.1.2
mmcv-full==1.4.3
multiprocess==0.70.14
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptflops==0.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.0
Shapely==1.8.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
thop==0.0.31.post2005241907
tifffile==2021.7.2
timm==0.4.12
torch==1.5.0
torch-tb-profiler==0.4.0
torchvision==0.6.0
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.2.3
wrapt==1.12.1
yapf==0.32.0
zipp==3.5.0
